pyproject.toml
[project]
name = "ultimate-planner"
version = "0.1.0"
description = "Ultimate Planner"
authors = [
  { name = "Chandima Ranaweera", email = "admin@tecbizsolutions.com" }
]
dependencies = [
  "fastapi",
  "uvicorn",
  "sqlalchemy>=2.0.0",
  "alembic",
  "psycopg2-binary",
  "aiosqlite",
  "python-dotenv",
  "pydantic[email]",
  "pytest",
  "pytest-asyncio",
  "jinja2>=3.1.6",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "python-multipart",
  "httpx",
  "cryptography",
  "pyotp",
  "qrcode[pil]",
  "pillow",
  "redis>=4.5.0",
  "aioredis>=2.0.0",
]
